+++ app/main.js
@@ -6,6 +6,7 @@ var loading_tmpl = require('./templates/loading.hbs');
 var error_tmpl = require('./templates/error.hbs');
 var socketio = require('socket.io-client');
 var errorNavigation = require('./error-navigation.js');
+var summarize = require('./summarize.js');
 require('./styles.less');
 require('bootstrap/dist/js/npm');
 
@@ -30,52 +31,31 @@ $(function () {
       .html(features_tmpl(tmpl_data))
       .find('[data-toggle="tooltip"]').tooltip();
 
-    var errors = [];
-    var summary = JSON.parse(JSON.stringify(event.plan));
-    summary.forEach(function(f, i){
-      f.scenarios.forEach(function(s, j){
-        s.status = 'pending';
-        if (i < event.snapshot.length){
-          if (event.snapshot[i].status === 'skipped'){
-            s.status = 'skipped';
-            return;
-          }
-          if (j < event.snapshot[i].elements.length){
-            s.status = 'passed';
-            event.snapshot[i].elements[j].steps.forEach(function(r){
-              if (r.result && r.result.status === 'failed'){
-                s.status = 'failed';
-                errors.push(s);
-              }
-            });
-          }
-        }
-      })
-    })
-  errorNavigation.register(errors);
+    var summaryResult = summarize(event);
+    errorNavigation.register(summaryResult.errors);
 
-  $('#summary').html(summary_tmpl({summary: summary}))
-      .find('[data-toggle="tooltip"]').tooltip();
-  });
+    $('#summary').html(summary_tmpl({summary: summaryResult.summary}))
+        .find('[data-toggle="tooltip"]').tooltip();
+    });
 
-  socket.on('tests_complete', function () {
-    $('#run-tests').prop('disabled', false);
-  });
-  socket.on('disconnect', function () {
-    $('#run-tests').prop('disabled', false);
-    $('#status').html(error_tmpl({
-      'responseText': 'Disconnected from server.'
-    }));
-  });
+    socket.on('tests_complete', function () {
+      $('#run-tests').prop('disabled', false);
+    });
+    socket.on('disconnect', function () {
+      $('#run-tests').prop('disabled', false);
+      $('#status').html(error_tmpl({
+        'responseText': 'Disconnected from server.'
+      }));
+    });
 
-  $('#run-tests').on('click', function (event) {
-    var vendor = $('#vendor').val();
-    errorNavigation.reset();
+    $('#run-tests').on('click', function (event) {
+      var vendor = $('#vendor').val();
+      errorNavigation.reset();
 
-    $(event.currentTarget).prop('disabled', true);
-    socket.emit('begin_tests', {vendor: vendor})
+      $(event.currentTarget).prop('disabled', true);
+      socket.emit('begin_tests', {vendor: vendor})
 
-    $('#summary').html("");
-    $('#canvas').html(loading_tmpl({}));
-  });
+      $('#summary').html("");
+      $('#canvas').html(loading_tmpl({}));
+    });
 });
